+++ mne/utils/docs.py
@@ -76,6 +76,35 @@ on_missing : str
         Behavior when ``stc`` has vertices that are not in ``fwd``.
         Can be "ignore", "warn"", or "raise"."""
 
+# Simulation
+docdict['random_state'] = """
+random_state : None | int | ~numpy.random.RandomState
+    The random generator state used for blink, ECG, and sensor
+    noise randomization. Default is None, which does not set the seed.
+"""
+docdict['interp'] = """
+interp : str
+    Either 'hann', 'cos2' (default), 'linear', or 'zero', the type of
+    forward-solution interpolation to use between forward solutions
+    at different head positions.
+"""
+docdict['head_pos'] = """
+head_pos : None | str | dict | tuple | array
+    Name of the position estimates file. Should be in the format of
+    the files produced by MaxFilter. If dict, keys should
+    be the time points and entries should be 4x4 ``dev_head_t``
+    matrices. If None, the original head position (from
+    ``info['dev_head_t']``) will be used. If tuple, should have the
+    same format as data returned by `head_pos_to_trans_rot_t`.
+    If array, should be of the form returned by
+    :func:`mne.chpi.read_head_pos`.
+"""
+docdict['n_jobs'] = """
+n_jobs : int
+    The number of jobs to run in parallel (default 1).
+    Requires the joblib package.
+"""
+
 # Finalize
 docdict = unindent_dict(docdict)
 fill_doc = filldoc(docdict, unindent_params=False)